// src/types.ts
export enum Weather {
  Sunny = 'sunny',
  Rainy = 'rainy',
  Cloudy = 'cloudy',
  Windy = 'windy',
  Stormy = 'stormy'
}

export enum Visibility {
  Great = 'great',
  Good = 'good',
  Ok = 'ok',
  Poor = 'poor'
}

export interface DiaryEntries {
  id: number
  date: string
  weather: Weather
  visibility: Visibility
  comment: string
}

export type DiarySensitiveEntries = Pick<DiaryEntries, 'id' | 'date' | 'weather' | 'visibility'>
export type NewDiaryEntry = Omit<DiaryEntries, 'id'>